Major changes to proxy and btp services.
Proxy rules now operate on "domains" at a top level. A domain
may include wildcards. Underneath domains are paths. A proxy
resource handler may redirect or proxy a path to a URL.

No more concept of "permitted hosts". This is taken care of
by specifying the domain that a service resides on at a lower
level now.

Added list of default domains to add btp services to. If no
domain is specified, a service will be added to all domains
in the node config. If no default domains are specified,
then "*" will be used for the default domain.